requirements.txt
pandas>=1.5.0
scikit-learn==1.6.1
joblib==1.4.2
numpy>=1.21.0
openai>=1.0.0
python-dotenv>=0.19.0
requests>=2.28.0
xlrd>=2.0.0
openpyxl>=3.0.0
# FastAPI e dependências
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.0.0

# PDF parsing
pdfplumber>=0.10.0
PyPDF2>=3.0.0
python-multipart>=0.0.6

# Testes
pytest>=7.0.0
